Use gsl::index for returning an index to a container item. Use gsl::narrow_cast for casting from double to float in D2D functions. Use auto to avoid conversions. Use default member initialization.

+++ src/msw/cursor.cpp
@@ -272,7 +272,7 @@ wxSize ScaleAndReverseBitmap(HBITMAP& bitmap, float scale)
     BITMAP bmp;
     if ( !::GetObject(bitmap, sizeof(bmp), &bmp) )
         return {};
-    wxSize cs(bmp.bmWidth * scale, bmp.bmHeight * scale);
+    wxSize cs(gsl::narrow_cast<int>(bmp.bmWidth * scale), gsl::narrow_cast<int>(bmp.bmHeight * scale));
 
     MemoryHDC hdc;
     SelectInHDC selBitmap(hdc, bitmap);
